requirements.txt
gunicorn==20.1.0
flask-cors==3.0.10
python-dotenv==1.0.0
supabase==2.3.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2